Revert "Merge branch 'ContactList' of https://github.com/Study-Buddies-SwEnt/SwEnt_Group_project into ContactList"
This reverts commit bbf1664e9b3b4e2cb1fb6edfe6b3a4b1fe62de99, reversing
changes made to 18571e75ea6744f8a40806ba6453c58d21a149bd.
