+++ app/src/androidTest/java/com/github/se/studybuddies/tests/ChatTest.kt
@@ -18,12 +18,12 @@ import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.screens.ChatScreen
 import com.github.se.studybuddies.ui.chat.ChatScreen
 import com.github.se.studybuddies.ui.chat.EditDialog
-import com.github.se.studybuddies.ui.chat.IconsOptionsList
 import com.github.se.studybuddies.ui.chat.MessageBubble
-import com.github.se.studybuddies.ui.chat.OptionsDialog
-import com.github.se.studybuddies.ui.chat.SendFileMessage
-import com.github.se.studybuddies.ui.chat.SendLinkMessage
-import com.github.se.studybuddies.ui.chat.SendPhotoMessage
+import com.github.se.studybuddies.ui.chat.utility.IconsOptionsList
+import com.github.se.studybuddies.ui.chat.utility.OptionsDialog
+import com.github.se.studybuddies.ui.chat.utility.SendFileMessage
+import com.github.se.studybuddies.ui.chat.utility.SendLinkMessage
+import com.github.se.studybuddies.ui.chat.utility.SendPhotoMessage
 import com.github.se.studybuddies.viewModels.MessageViewModel
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
