+++ app/src/main/java/com/github/se/studybuddies/ui/shared_elements/SharedUI.kt
@@ -262,6 +262,16 @@ fun GoBackRouteButton(
 }
 
 @Composable
+fun GoBackRouteToLastPageButton(
+    navigationActions: NavigationActions,
+) {
+  Icon(
+      imageVector = Icons.Default.ArrowBack,
+      contentDescription = "Go back",
+      modifier = Modifier.clickable { navigationActions.goBack() }.testTag("go_back_button"))
+}
+
+@Composable
 fun GroupsSettingsButton(navigationActions: NavigationActions) {
   val expandedState = remember { mutableStateOf(false) }
   IconButton(
