+++ app/src/main/java/com/github/se/studybuddies/viewModels/GroupViewModel.kt
@@ -81,8 +81,29 @@ class GroupViewModel(
     viewModelScope.launch { db.deleteGroup(groupUID) }
   }
 
-  fun addUserToGroup(groupUID: String, text: String = "") {
-    viewModelScope.launch { db.addUserToGroup(groupUID, text) }
+  fun addUserToGroup(groupUID: String, text: String = "", callBack: (Boolean) -> Unit) {
+    viewModelScope.launch {
+      db.addUserToGroup(groupUID, text) { isError -> callBack(isError) }
+      val userUID: String =
+          if (text != "") {
+            text
+          } else {
+            db.getCurrentUser().toString()
+          }
+      val newUser = db.getUser(userUID)
+      if (newUser != User.empty()) {
+        var updatedMembers = _members.value?.toMutableList()
+        if (newUser != null) {
+          updatedMembers?.add(newUser)
+        }
+        updatedMembers = updatedMembers?.toSet()?.toMutableList()
+        _members.value = updatedMembers ?: emptyList()
+      }
+    }
+  }
+
+  fun addSelfToGroup(groupUID: String) {
+    viewModelScope.launch { db.addSelfToGroup(groupUID) }
   }
 
   fun updateGroup(groupUID: String, name: String, photoURI: Uri?) {
