+++ app/src/main/res/values/strings.xml
@@ -140,6 +140,7 @@
     <string name="user_profile_picture">user profile picture</string>
     <string name="delete_topic">Delete Topic</string>
     <string name="loading">loading</string>
+    <string name="loadingP">loading…</string>
     <string name="add_member_from_list">Add member from list</string>
     <string name="add_member_with_uid">Add member with UID</string>
     <string name="joining">Joining …</string>
@@ -147,6 +148,7 @@
     <string name="share_location">Share location</string>
     <string name="location_shared">Location is shared</string>
     <string name="location_not_shared">Location is not shared</string>
+    <string name="add_member">Add member</string>
     <string name="enter_poll_question">Enter poll question</string>
     <string name="poll_question">Question</string>
     <string name="poll_options">Options</string>
